separate opt_model and test_cache
